Rename project to GenerateCppFiltersExtension and align files and folders accordingly
