+++ GenerateCppFiltersExtension.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26228.9
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenerateFilter", "GenerateFilter\GenerateFilter.csproj", "{18555A86-3A5F-4D33-8C12-01E288BDC30E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenerateCppFiltersExtension", "GenerateCppFiltersExtension\GenerateCppFiltersExtension.csproj", "{18555A86-3A5F-4D33-8C12-01E288BDC30E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
