+++ GenerateCppFiltersExtension/Resources/GenerateCppFilters.ico
