+++ GenerateCppFiltersExtension/Resources/GenerateCppFiltersPackage.ico
