+++ GenerateCppFiltersExtension/Resources/PreviewImage.png
